Fix breaking change in freetype 2.5.5
From here:
http://git.savannah.gnu.org/cgit/freetype/freetype2.git/commit/?id=b3500af717010137046ec4076d1e1c0641e33727

+++ bitmap.go
@@ -44,7 +44,7 @@ func (b *Bitmap) Pitch() int {
 
 // Buffer returns the bitmap buffer.
 func (b *Bitmap) Buffer() []byte {
-	l := b.handle.rows * b.handle.pitch
+	l := C.int(b.handle.rows) * b.handle.pitch
 	return C.GoBytes(unsafe.Pointer(b.handle.buffer), l)
 }
